+++ src/commands/forceskip.js
@@ -12,13 +12,21 @@ exports.run = function (client, msg, args, options, sel) {
 	}});
 
 	client.voiceConnections.get(msg.guild.id).dispatcher.end();
+	let skip = /^\d+$/.test(args) ? args : sel;
+	if (skip) {
+		client.queues[msg.guild.id].queue.splice(0, skip-1)
+		msg.channel.send({ embed: {
+			color: config.options.embedColour,
+			title: `Skipped ${skip} songs`
+		}})
+	}
 
 }
 
 exports.usage = {
 	main: "{prefix}{command}",
-	args: "",
-	description: "Skips the current song by force (no voting)",
+	args: "[number]",
+	description: "Skips the current song by force (no voting)\nCan also skip multiple songs by adding [number]. e.g. `.forceskip 5`",
 	adminOnly: true,
 	DJ: true
 };